refactor : refactor error package
