+++ server/src/main/java/tig/server/global/response/ApiResponse.java
@@ -1,9 +1,7 @@
-package tig.server.error;
+package tig.server.global.response;
 
-import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 /**
  * [공통] API Response 결과의 반환 값을 관리
