+++ server/src/main/java/tig/server/global/response/ErrorResponse.java
@@ -1,10 +1,11 @@
-package tig.server.error;
+package tig.server.global.response;
 
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.validation.BindingResult;
+import tig.server.global.code.ErrorCode;
 
 import java.util.ArrayList;
 import java.util.List;
