+++ server/src/main/java/tig/server/jwt/handler/JwtExceptionHandlerFilter.java
@@ -16,8 +16,6 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
-import tig.server.error.BusinessExceptionHandler;
-import tig.server.error.ErrorCode;
 import tig.server.jwt.TokenProvider;
 import tig.server.member.service.MemberDetailsServiceImpl;
