+++ server/src/main/java/tig/server/member/service/MemberService.java
@@ -4,14 +4,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import tig.server.enums.MemberRoleEnum;
-import tig.server.error.BusinessExceptionHandler;
-import tig.server.error.ErrorCode;
+import tig.server.global.exception.BusinessExceptionHandler;
+import tig.server.global.code.ErrorCode;
 import tig.server.jwt.TokenProvider;
 import tig.server.kakao.dto.KakaoUserInfoResponseDto;
 import tig.server.kakao.dto.LoginMemberResponseDto;
 import tig.server.member.domain.Member;
 import tig.server.member.dto.MemberResponse;
-import tig.server.member.dto.RefreshTokenRequestDto;
 import tig.server.member.dto.RefreshTokenResponseDto;
 import tig.server.member.mapper.MemberMapper;
 import tig.server.member.repository.MemberRepository;
