+++ server/src/main/java/tig/server/payment/service/PaymentService.java
@@ -14,8 +14,8 @@ import reactor.core.publisher.Mono;
 import tig.server.club.domain.Club;
 import tig.server.club.repository.ClubRepository;
 import tig.server.enums.Type;
-import tig.server.error.BusinessExceptionHandler;
-import tig.server.error.ErrorCode;
+import tig.server.global.exception.BusinessExceptionHandler;
+import tig.server.global.code.ErrorCode;
 import tig.server.payment.dto.PaymentCompleteResponseDto;
 import tig.server.payment.dto.PaymentRequestDto;
 import tig.server.payment.dto.PaymentResponseDto;
