+++ server/src/main/java/tig/server/reservation/service/ReservationService.java
@@ -10,8 +10,8 @@ import tig.server.discord.DiscordMessageProvider;
 import tig.server.discord.EventMessage;
 import tig.server.enums.Status;
 import tig.server.enums.Type;
-import tig.server.error.BusinessExceptionHandler;
-import tig.server.error.ErrorCode;
+import tig.server.global.exception.BusinessExceptionHandler;
+import tig.server.global.code.ErrorCode;
 import tig.server.member.domain.Member;
 import tig.server.member.mapper.MemberMapper;
 import tig.server.member.service.MemberService;
