+++ server/src/main/java/tig/server/search/controller/SearchController.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import tig.server.annotation.LoginUser;
-import tig.server.error.ApiResponse;
+import tig.server.global.response.ApiResponse;
 import tig.server.member.domain.Member;
 import tig.server.search.dto.SearchResultDto;
 import tig.server.search.service.SearchService;
