+++ server/src/main/java/tig/server/wishlist/service/WishlistService.java
@@ -7,8 +7,8 @@ import tig.server.club.domain.Club;
 import tig.server.club.dto.ClubResponse;
 import tig.server.club.mapper.ClubMapper;
 import tig.server.club.service.ClubService;
-import tig.server.error.BusinessExceptionHandler;
-import tig.server.error.ErrorCode;
+import tig.server.global.exception.BusinessExceptionHandler;
+import tig.server.global.code.ErrorCode;
 import tig.server.member.domain.Member;
 import tig.server.member.dto.MemberResponse;
 import tig.server.member.mapper.MemberMapper;
